backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI-powered customer support agent"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.13.0",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.20.0",
    "fastapi[standard]>=0.129.0",
    "google-generativeai>=0.8.6",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "openai-agents>=0.9.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=1.40.0",
    "sendgrid>=6.11.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=0.21.0",
]
